IT'S ALIVE

+++ xtrek/src/xtrek/OnGPSUpdateListener.java
@@ -3,8 +3,8 @@ package xtrek;
 import java.util.EventListener;
 
 interface OnGPSUpdateListener extends EventListener {
-    void onGPSUpdate(Float latitude,
-                     Float longitude,
+    void onGPSUpdate(Double latitude,
+                     Double longitude,
                      SatelliteModel.Direction latitudeDirection,
                      SatelliteModel.Direction longitudeDirection);
 }
