+++ xtrek/src/xtrek/Satellite.java
@@ -75,9 +75,9 @@ public class Satellite extends Mode {
 
     private void updatePosition() {
         //TODO check that values are not empty to determine mad bants
-        Float value1 = SModel.getLatitude();
+        Double value1 = SModel.getLatitude();
         String direction1 = SModel.getLatitudeDirection().getDirection();
-        Float value2 = SModel.getLongitude();
+        Double value2 = SModel.getLongitude();
         String direction2 = SModel.getLongitudeDirection().getDirection();
 
         if (value1 == null) {
