+++ xtrek/src/xtrek/SatelliteView.java
@@ -18,8 +18,8 @@ public class SatelliteView extends ModeView {
         super(frame);
         panel.setLayout(new GridBagLayout());
     }
-    
-    public void setPosition(Float value1, String direction1, Float value2, String direction2) {
+
+    public void setPosition(Double value1, String direction1, Double value2, String direction2) {
         sat.setLabel1(value1 + " " + direction1);
         sat.setLabel2(value2 + " " + direction2);
     }
